Fix a circular dep issue that Python doesn't like (surprised, Node has no issues with this) by moving code from parse to meta. But most of those functions weren't used by anything else so this is a net positive
